修改AppCrawlerConfig表的websiteType字段类型

+++ rawdatacentral-domain/src/main/java/com/datatrees/rawdatacentral/domain/appconfig/CrawlerProjectParam.java
@@ -13,17 +13,17 @@ public class CrawlerProjectParam {
     /**
      * 业务类型 {WebsiteType}
      */
-    private Integer            websiteType;
+    private Byte            websiteType;
     /**
      * 具体业务
      */
     private List<ProjectParam> projects;
 
-    public Integer getWebsiteType() {
+    public Byte getWebsiteType() {
         return websiteType;
     }
 
-    public void setWebsiteType(Integer websiteType) {
+    public void setWebsiteType(Byte websiteType) {
         this.websiteType = websiteType;
     }
